refactor(win): move render logic out from window content
