+++ src/ui/tree.rs
@@ -6,7 +6,7 @@ use crate::cart::{IRect, U16Rect, U16Size};
 use crate::glovar;
 use crate::ui::canvas::{Canvas, CanvasArc};
 use crate::ui::tree::internal::{InodeId, Inodeable, Itree, ItreeIter, ItreeIterMut};
-use crate::ui::widget::window::WindowOptions;
+use crate::ui::widget::window::WindowLocalOptions;
 use crate::ui::widget::{Cursor, RootContainer, Widgetable, Window};
 
 use parking_lot::RwLock;
@@ -113,14 +113,15 @@ impl Widgetable for TreeNode {
 #[derive(Debug, Clone)]
 /// Global options for UI.
 pub struct GlobalOptions {
-  /// Window options.
-  pub window_options: WindowOptions,
+  /// Window local options.
+  pub window_local_options: WindowLocalOptions,
 }
 
 impl Default for GlobalOptions {
   fn default() -> Self {
-    let window_options = WindowOptions::builder().build();
-    GlobalOptions { window_options }
+    GlobalOptions {
+      window_local_options: WindowLocalOptions::builder().build(),
+    }
   }
 }
 
@@ -441,31 +442,31 @@ impl Tree {
   }
 
   pub fn wrap(&self) -> bool {
-    self.options.window_options.wrap()
+    self.options.window_local_options.wrap()
   }
 
   pub fn set_wrap(&mut self, value: bool) {
-    self.options.window_options.set_wrap(value);
+    self.options.window_local_options.set_wrap(value);
   }
 
   pub fn line_break(&self) -> bool {
-    self.options.window_options.line_break()
+    self.options.window_local_options.line_break()
   }
 
   pub fn set_line_break(&mut self, value: bool) {
-    self.options.window_options.set_line_break(value);
+    self.options.window_local_options.set_line_break(value);
   }
 
   pub fn breat_at(&self) -> &String {
-    self.options.window_options.break_at()
+    self.options.window_local_options.break_at()
   }
 
   pub fn set_break_at(&mut self, value: &str) {
-    self.options.window_options.set_break_at(value);
+    self.options.window_local_options.set_break_at(value);
   }
 
   pub fn break_at_regex(&self) -> &Regex {
-    self.options.window_options.break_at_regex()
+    self.options.window_local_options.break_at_regex()
   }
 
   // Global options }
